backend-auth/requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
pyjwt
psycopg2-binary
